Chapter 4: tables and examples (invented two example roles).
The first invented role is role="lojbanization-example", for lojbanization of jargon (i.e. fu'ivla making) or names.

The second is role="lujvo-making-example", and it probably only applies to a certain four examples which illustrate the lujvo-making algorithm. Here, <rafsi> encloses the rafsi used to make the lujvo and <score> encloses the equation of the luvjo's score.