Chapter 2: <example>s. Invented several tags and attributes.
<sumti>, <selbri>, <cmavo>, <elidable> for the interlinear
glosses. <elidable> can have elidable="false", which means it's
wrapping an elidable terminator that cannot be elided, or
delineated="false", which means it's wrapping an elidable terminator
that can be elided but for some reason wasn't surrounded with []
(there's only of these). <cmavo> is for UI and such.

Also, <place-structure> now has to be wrapped in <para> (i.e. it is no
longer a child of the same parents that <para> can be a child of)