Update the web version of CLL (Issue #239).
Hopefully my changes *only* affect the web styles, and not print... I was mainly
concerned with adjusting font size and whitespace to improve legibility, and
making the layout more responsive so it can adjust itself to fit the user's
display device.

I moved all the `@media screen` styles to the end of the stylesheet, to make it
easier to override the print styles.

Font sizes make more sense (on the web) when using CSS pixels, since other units --
even ones that you'd expect to correspond to a fixed physical length like inches
and centimeters -- actually get converted to this "average" pixel size anyway,
when rendered in a browser. An inch or a point may not actually be as big as you
think it should be, depending on the user's display device.

I have font sizes set based on the display size, because users tend to hold
small devices closer to their face, while very large displays tend to be mounted
further away from the user's seat.