SO MANY CHANGES.
We now build our own EPUB instead of using calibre conversion.

From the changelog:

The HTML version passes tidy without errors (although it does have
some bizarre "<a> anchor [blah] already defined" warnings that
appear to simply be bugs in tidy).

This was the "fix up the EPUB" version; we now roll our own EPUB,
and it passes EPubCheck 4.0.1 without errors.

As a side effect of the EPUB work, the html was cleaned up
considerably, but not in any ways that should have any visual
effects; there are now no nested "a" elements, all "dt" have "dd".

The links to the glossary now work (they were there but invisible
before).  The HTML and EPUB indexes now us ":" between items and
linuks (before it was a comma, which just looks weird).