I'll just quote from the changelog:
2016-08-26
==========

Fix the EPUB (and the Kindle MOBI) round two!

All versions:
- A couple of insignificant wording changes; https://github.com/lojban/cll/commit/6b891e06ec0115877fd062b3c141b7c6ec3060c8
- The definition of bakrecpa'o in jbovlaste changed.  :)

EPUB generation was substantially altered to make kindlegen happy.
In particular
- tags that had previously been ignored with "display: none;" are
  now removed entirely
- dl/dd/dt tags in the EPUB are now replace with ul/li tags with
  special CSS

Also, there is now an official MOBI, the one generated by kindlegen
from the EPUB.

This version was only pushed to the EPUB on the publisher side; the
changes to the PDF were so minor that it didn't seem worth it to
push the print version as well.