.i zvati la iupiter .i na zvati la iupiter
Section 19, example 19.4 is translated as "I opine the fact-that a-mass-of living-things is-at Jupiter or-else I opine the fact-that a-mass-of living-things isn't-at Jupiter" but really it's "I opine the fact-that a-mass-of living-things is-at Jupiter or-else I do-not opine the fact-that a-mass-of living-things is-at Jupiter". So 19.4 is actually true, not false as it is claimed to be.

Correct, so this example will need replacing. John Cowan: Approved