chapters: Un-forbid the -mz- consonant cluster.
ti'e ka'u This is only forbidden as an insult to James Cooke Brown.
However, this insult no longer makes sense to sustain, as it can no
longer be delivered as intended. Instead, it now harms anybody who might
desire the -mz- cluster for whatever use, including for Lojbanization of
existing names.

Frankly, I can't believe I have to argue in favor of this. It's obvious.